requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
jinja2>=3.1.6
python-multipart>=0.0.20
Pillow>=11.3.0
pillow-heif>=1.3.0
rawpy>=0.26.1
PyMuPDF>=1.27.2
pytest>=8.4.2
httpx>=0.28.1
